refactor: stats 컴포넌트 타입 @/types로 이동
